refactor: remove storage core compat bridge
