+++ docs/architecture/migration-progress.md
@@ -5,16 +5,16 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 ## Current Context
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
-- Branch: `overtrue/arch-app-admin-secondary-compat-bridges`
-- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116`.
-- Based on: API-116 slice.
+- Branch: `overtrue/arch-storage-core-secondary-compat-bridge`
+- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117`.
+- Based on: API-117 slice.
 - PR type for this branch: `pure-move`
 - Runtime behavior changes: none.
-- Rust code changes: collapse app/admin secondary compatibility bridge consumers
-  directly into owner `storage_compat` boundaries.
-- CI/script changes: guard against reintroducing app/admin secondary
-  compatibility bridge modules.
-- Docs changes: record the API-117 app/admin secondary bridge cleanup.
+- Rust code changes: collapse storage owner secondary compatibility bridge
+  consumers directly into `storage_compat`.
+- CI/script changes: guard against reintroducing the storage
+  `core_storage_compat` bridge module.
+- Docs changes: record the API-118 storage secondary bridge cleanup.
 
 ## Phase 0 Tasks
 
@@ -659,6 +659,18 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
   - Verification: RustFS compile coverage, app/admin secondary bridge residual
     scan, migration and layer guards, formatting, diff hygiene, Rust risk scan,
     and three-expert review.
+- [x] `API-118` Remove storage core secondary compatibility bridge.
+  - Completed slice: replace storage owner consumers of `core_storage_compat`
+    with direct `storage_compat` paths, then delete the secondary bridge module.
+  - Acceptance: storage owner modules and tests no longer route through a
+    second local compatibility bridge; migration rules reject reintroduced
+    `core_storage_compat` references.
+  - Must preserve: ECFS object operations, storage access checks, SSE
+    encryption helpers, storage option resolution, and existing owner
+    `storage_compat` aliases.
+  - Verification: RustFS compile coverage, storage secondary bridge residual
+    scan, migration and layer guards, formatting, diff hygiene, path-only risk
+    review, and three-expert review.
 - [x] `G-012` Inventory placement and repair invariants.
   - Acceptance:
     [`placement-repair-invariants.md`](placement-repair-invariants.md) records
@@ -3692,13 +3704,24 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 
 | Expert | Status | Notes |
 |---|---|---|
-| Quality/architecture | pass | API-117 removes app/admin secondary compatibility bridges while keeping consumers on their owner `storage_compat` boundaries. |
-| Migration preservation | pass | The extended guard rejects reintroduced `router_storage_compat` and `usecase_storage_compat` bridge names while preserving existing owner aliases. |
-| Testing/verification | pass | RustFS compile, app/admin secondary bridge residual scan, migration guard, layer guard, formatting, diff hygiene, and path-only risk review passed. |
+| Quality/architecture | pass | API-118 removes the storage core secondary compatibility bridge while keeping consumers on the storage owner `storage_compat` boundary. |
+| Migration preservation | pass | The extended guard rejects reintroduced `core_storage_compat` references while preserving existing owner aliases. |
+| Testing/verification | pass | RustFS compile, storage secondary bridge residual scan, migration guard, layer guard, formatting, diff hygiene, and path-only risk review passed. |
 
 ## Verification Notes
 
 Passed before push:
+
+- Issue #660 API-118 current slice:
+  - `cargo check -p rustfs --tests`: passed.
+  - `cargo fmt --all`: passed.
+  - `cargo fmt --all --check`: passed.
+  - `git diff --check`: passed.
+  - `bash -n scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_layer_dependencies.sh`: passed.
+  - Storage secondary compatibility bridge residual scan: passed.
+  - Rust risk review on path-only replacements and guard script: passed.
 
 - Issue #660 API-117 current slice:
   - `cargo check -p rustfs --tests`: passed.
