+++ rustfs/src/storage/mod.rs
@@ -15,7 +15,6 @@
 pub mod access;
 pub mod backpressure;
 pub mod concurrency;
-pub(crate) mod core_storage_compat;
 pub mod deadlock_detector;
 pub mod ecfs;
 pub(crate) mod helper;
@@ -30,11 +29,11 @@ pub mod timeout_wrapper;
 pub mod tonic_service;
 
 pub(crate) type StorageDeletedObject = rustfs_storage_api::DeletedObject;
-pub(crate) type StorageGetObjectReader = self::core_storage_compat::GetObjectReader;
-pub(crate) type StorageObjectInfo = self::core_storage_compat::ObjectInfo;
-pub(crate) type StorageObjectOptions = self::core_storage_compat::ObjectOptions;
+pub(crate) type StorageGetObjectReader = self::storage_compat::GetObjectReader;
+pub(crate) type StorageObjectInfo = self::storage_compat::ObjectInfo;
+pub(crate) type StorageObjectOptions = self::storage_compat::ObjectOptions;
 pub(crate) type StorageObjectToDelete = rustfs_storage_api::ObjectToDelete;
-pub(crate) type StoragePutObjReader = self::core_storage_compat::PutObjReader;
+pub(crate) type StoragePutObjReader = self::storage_compat::PutObjReader;
 
 #[cfg(test)]
 mod concurrent_fix_test;
